Reduce main file size by extracting tests
